Split DeepCachingType into its own source file
