+++ lib/go-tc/enum/enum.go
@@ -29,8 +29,6 @@ package enum
  */
 
 import (
-	"encoding/json"
-	"errors"
 	"strconv"
 	"strings"
 )
@@ -125,64 +123,6 @@ func DSTypeCategoryFromString(s string) DSTypeCategory {
 
 const SigningAlgorithmURLSig = "url_sig"
 const SigningAlgorithmURISigning = "uri_signing"
-
-// DeepCachingType represents a Delivery Service's deep caching type. The string values of this type should match the Traffic Ops values.
-type DeepCachingType string
-
-const (
-	DeepCachingTypeNever   = DeepCachingType("") // default value
-	DeepCachingTypeAlways  = DeepCachingType("ALWAYS")
-	DeepCachingTypeInvalid = DeepCachingType("INVALID")
-)
-
-// String returns a string representation of this deep caching type
-func (t DeepCachingType) String() string {
-	switch t {
-	case DeepCachingTypeAlways:
-		return string(t)
-	case DeepCachingTypeNever:
-		return "NEVER"
-	default:
-		return "INVALID"
-	}
-}
-
-// DeepCachingTypeFromString returns a DeepCachingType from its string representation, or DeepCachingTypeInvalid if the string is not a valid type.
-func DeepCachingTypeFromString(s string) DeepCachingType {
-	switch strings.ToLower(s) {
-	case "always":
-		return DeepCachingTypeAlways
-	case "never":
-		return DeepCachingTypeNever
-	case "":
-		// default when omitted
-		return DeepCachingTypeNever
-	default:
-		return DeepCachingTypeInvalid
-	}
-}
-
-// UnmarshalJSON unmarshals a JSON representation of a DeepCachingType (i.e. a string) or returns an error if the DeepCachingType is invalid
-func (t *DeepCachingType) UnmarshalJSON(data []byte) error {
-	if string(data) == "null" {
-		*t = DeepCachingTypeNever
-		return nil
-	}
-	s, err := strconv.Unquote(string(data))
-	if err != nil {
-		return errors.New(string(data) + " JSON not quoted")
-	}
-	*t = DeepCachingTypeFromString(s)
-	if *t == DeepCachingTypeInvalid {
-		return errors.New(string(data) + " is not a DeepCachingType")
-	}
-	return nil
-}
-
-// MarshalJSON marshals into a JSON representation
-func (t DeepCachingType) MarshalJSON() ([]byte, error) {
-	return json.Marshal(t.String())
-}
 
 type SteeringType string
